Issue#39 (#47)
* colocado bd no dm

* enhanc: adicionado sembast nas entidades

+++ in_a_bottle/lib/_shared/location/point.dart
@@ -19,7 +19,7 @@ class Point extends Equatable {
         point.latitude,
         point.longitude,
       ),
-    );
+    ).toInt();
   }
 
   Point copyWith({
